vector calc

+++ Controllers/NeuralController.cs
@@ -8,7 +8,7 @@ namespace NeuralNetworkServer.Controllers;
 public class NeuralController : ControllerBase
 {
 
-    private static readonly ConcurrentDictionary<long, (double[] data, int[] indexes)> InputsCache = new();
+    private static readonly ConcurrentDictionary<long, (float[] data, int[] indexes)> InputsCache = new();
 
     [HttpPost("Load")]
     public void Load(LoadInputsRequest request)
@@ -34,14 +34,14 @@ public class NeuralController : ControllerBase
         var inputKey = binaryReader.ReadInt64();
         
         var synapseLength = binaryReader.ReadInt32();
-        var synapses = new double[synapseLength];
+        var synapses = new float[synapseLength];
         for (var i = 0; i < synapseLength; i++)
         {
-            synapses[i] = binaryReader.ReadDouble();
+            synapses[i] = (float)binaryReader.ReadDouble();
         }
         var inputs = InputsCache[inputKey];
         var calcer = new NeuralNetworkCalcer(model);
-        var outputs = new double[inputs.indexes.Length][];
+        var outputs = new float[inputs.indexes.Length][];
         var outputsPerTask = 1000;
         var batchCount = inputs.indexes.Length / outputsPerTask;
         var tasks = new List<Task>();
