+++ Dto/CalcDto.cs
@@ -2,5 +2,5 @@
 public class CalcResponse
 {
     public List<long> UnkownInputs { get; set; }
-    public double[][] Outputs { get; set; }
+    public float[][] Outputs { get; set; }
 }
