+++ Dto/LoadInputsDto.cs
@@ -1,6 +1,6 @@
 public class LoadInputsRequest
 {
     public long Key { get; set; }
-    public double[] Data { get; set; }
+    public float[] Data { get; set; }
     public int[] Indexes { get; set; }
 }
